add games list on /games, moved invites to home, fix profile card width
